add generate_fee_string

+++ db/migrate/20240110232751_create_fees.rb
@@ -8,7 +8,7 @@ class CreateFees < ActiveRecord::Migration[7.1]
       t.string :academic_year, null: false
       t.text :installments, array: true, default: []
       t.integer :installment_num, default: 0
-      t.float :total_fee_paid, default: 0.0
+      t.integer :total_fee_paid, default: 0
       t.boolean :is_completed, default: false
 
       t.timestamps
